Allow for setting the hashing algorithm
This change allows the `manifest!` method to accept and
`algo:` paramter that indicates what kind of hashing algorithm
you want to use for the bag. If that paramter is not supplied
it uses md5 and sha1.

You can choose from md5, sha1, and sha256.

From the command line you can supply the hashing algo as well:

```sh
bagit add -f LICENSE.txt -a 'sha256' test
```