Initial rubocop formatting
This commit includes formatting changes
needed when running `rubocop` with the
`bixby` rules. The `.rubocop.yml` has
additional changes to the default rules.

It should get closer to the `bixby` rules
over time.

This also includes a change to the bag info
file so that it points to the current
GitHub repo and not the rubyforge address.